# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for crane manufacturing corporate website
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Spec {
  key: string;
  value: string;
}

export type ProductStatus = (typeof ProductStatus)[keyof typeof ProductStatus];

export const ProductStatus = {
  published: "published",
  draft: "draft",
} as const;

export interface Product {
  id: number;
  name: string;
  slug: string;
  categoryId: number;
  categoryName: string;
  shortDescription: string;
  description: string;
  specs: Spec[];
  coverImage: string;
  gallery: string[];
  pdfUrl?: string | null;
  status: ProductStatus;
  featured: boolean;
  sortOrder: number;
  capacity: string;
  usageAreas: string[];
  optionalEquipment: string[];
  createdAt: string;
  updatedAt: string;
}

export type CreateProductBodyStatus =
  (typeof CreateProductBodyStatus)[keyof typeof CreateProductBodyStatus];

export const CreateProductBodyStatus = {
  published: "published",
  draft: "draft",
} as const;

export interface CreateProductBody {
  name: string;
  slug: string;
  categoryId: number;
  shortDescription: string;
  description: string;
  specs: Spec[];
  coverImage: string;
  gallery: string[];
  pdfUrl?: string | null;
  status: CreateProductBodyStatus;
  featured: boolean;
  sortOrder: number;
  capacity: string;
  usageAreas: string[];
  optionalEquipment: string[];
}

export type UpdateProductBodyStatus =
  (typeof UpdateProductBodyStatus)[keyof typeof UpdateProductBodyStatus];

export const UpdateProductBodyStatus = {
  published: "published",
  draft: "draft",
} as const;

export interface UpdateProductBody {
  name?: string;
  categoryId?: number;
  shortDescription?: string;
  description?: string;
  specs?: Spec[];
  coverImage?: string;
  gallery?: string[];
  pdfUrl?: string | null;
  status?: UpdateProductBodyStatus;
  featured?: boolean;
  sortOrder?: number;
  capacity?: string;
  usageAreas?: string[];
  optionalEquipment?: string[];
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  image: string;
  sortOrder: number;
  productCount: number;
}

export interface CreateCategoryBody {
  name: string;
  slug: string;
  description: string;
  image: string;
  sortOrder: number;
}

export type NewsArticleStatus =
  (typeof NewsArticleStatus)[keyof typeof NewsArticleStatus];

export const NewsArticleStatus = {
  published: "published",
  draft: "draft",
} as const;

export interface NewsArticle {
  id: number;
  title: string;
  slug: string;
  summary: string;
  content: string;
  image: string;
  date: string;
  status: NewsArticleStatus;
  createdAt: string;
}

export type CreateNewsBodyStatus =
  (typeof CreateNewsBodyStatus)[keyof typeof CreateNewsBodyStatus];

export const CreateNewsBodyStatus = {
  published: "published",
  draft: "draft",
} as const;

export interface CreateNewsBody {
  title: string;
  slug: string;
  summary: string;
  content: string;
  image: string;
  date: string;
  status: CreateNewsBodyStatus;
}

export type MediaItemType = (typeof MediaItemType)[keyof typeof MediaItemType];

export const MediaItemType = {
  video: "video",
  gallery: "gallery",
  pdf: "pdf",
} as const;

export interface MediaItem {
  id: number;
  type: MediaItemType;
  title: string;
  url: string;
  thumbnail?: string | null;
  createdAt: string;
}

export type CreateMediaBodyType =
  (typeof CreateMediaBodyType)[keyof typeof CreateMediaBodyType];

export const CreateMediaBodyType = {
  video: "video",
  gallery: "gallery",
  pdf: "pdf",
} as const;

export interface CreateMediaBody {
  type: CreateMediaBodyType;
  title: string;
  url: string;
  thumbnail?: string | null;
}

export interface ContactFormBody {
  name: string;
  email: string;
  phone?: string | null;
  company?: string | null;
  department?: string | null;
  subject: string;
  message: string;
}

export interface ContactFormResponse {
  success: boolean;
  message: string;
}

export interface CompanyStats {
  yearsOfExperience: number;
  productsDelivered: number;
  countriesServed: number;
  certifications: number;
  productCount: number;
  categoryCount: number;
}

export type ListProductsParams = {
  categoryId?: number;
  status?: ListProductsStatus;
  featured?: boolean;
};

export type ListProductsStatus =
  (typeof ListProductsStatus)[keyof typeof ListProductsStatus];

export const ListProductsStatus = {
  published: "published",
  draft: "draft",
} as const;

export type ListNewsParams = {
  status?: ListNewsStatus;
  limit?: number;
};

export type ListNewsStatus =
  (typeof ListNewsStatus)[keyof typeof ListNewsStatus];

export const ListNewsStatus = {
  published: "published",
  draft: "draft",
} as const;

export type ListMediaParams = {
  type?: ListMediaType;
};

export type ListMediaType = (typeof ListMediaType)[keyof typeof ListMediaType];

export const ListMediaType = {
  video: "video",
  gallery: "gallery",
  pdf: "pdf",
} as const;