# lib/api-zod/src/generated/types/companyStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for crane manufacturing corporate website
 * OpenAPI spec version: 0.1.0
 */

export interface CompanyStats {
  yearsOfExperience: number;
  productsDelivered: number;
  countriesServed: number;
  certifications: number;
  productCount: number;
  categoryCount: number;
}