# lib/api-zod/src/generated/types/contactFormBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for crane manufacturing corporate website
 * OpenAPI spec version: 0.1.0
 */

export interface ContactFormBody {
  name: string;
  email: string;
  phone?: string | null;
  company?: string | null;
  department?: string | null;
  subject: string;
  message: string;
}